Added temporary textures and prefabs for doctor, defibulator, and patient.
The doctor can now pick up a tool by pressing space (with the temp
movement script). This will set thisDoc.currentTool to the instance
of the defibulator (of base class Tool), and make the tool a child
object of the doctor object. Missing many things, like range checking
-- simple proof of concept. This does demonstrate the route I wish
to go with doctor-tool relationships. The same tools should persist
throughout the levels. We don't want tools to be reinstantiated when
they are dropped, or remove them from the level in some way when they
are picked up. Preserving references will prevent reinstantiation.